test: extract tbclient mock.
